Support Deployer Notifications (#79)
Part of work for https://ucsc-cgl.atlassian.net/browse/SEAB-2276

The deployer will publish messages to SNS (work in progress in dockstore-deploy). The deployer will set the source to dockstore.deployer.

This PR enables the lambda to process such events.

As part of this, broke the code up into several functions to make it more readable, at least for me.

Also added linting as for this one lambda. There are other lambdas in the repo, but I was only working
on this one. One step at a time.